Add WaitForJobCompleted for RDSv3
